requirements.txt
# Multi-Agent Analytics System - Dependencies

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# LLM Client
openai==1.10.0

# HTTP Client (pinned for openai compatibility)
httpx>=0.25.0,<0.28.0

# Google APIs
google-analytics-data==0.18.5
google-auth==2.27.0
gspread==5.12.3

# Data Processing
pandas==2.2.0

# Date/Time Handling
python-dateutil==2.8.2

# HTTP Client
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0